refactor(views): share album grids
