+++ crates/views/src/shared/mod.rs
@@ -1,11 +1,10 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 pub(crate) mod adaptive;
-pub(crate) mod card_grid;
+pub(crate) mod album_grid;
 pub(crate) mod cells;
 pub(crate) mod hero;
 pub(crate) mod menu;
 pub(crate) mod page;
 pub(crate) mod playlist_editor;
-pub(crate) mod release_card;
 pub(crate) mod tracks;
